% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_cats}
\alias{collapse_cats}
\title{Collapses factor levels and returns a new data frame}
\usage{
collapse_cats(data, item, collaps_cats)
}
\arguments{
\item{data}{dataset}

\item{item}{factor variable}

\item{collaps_cats}{list of factor levels to collapse in shape of \code{list(new_level = old_level)}}
}
\value{
new data frame with collapsed factor levels
}
\description{
This function is a wrapper around \code{forcats::fct_collapse()}.
It takes a dataframe, a factor variable, and a list of factor levels to collapse.
}
\examples{
collapse_cats(trust, police, list(Agree = c("Definitely Agree", "Rather Agree")))
}
