% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-template.R
\name{create_template}
\alias{create_template}
\title{This package will create a function called create_template()}
\usage{
create_template(path, ...)
}
\description{
It's callback is at: inst/rstudio/templates/project/create_template.dcf
}
