% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{desub}
\alias{desub}
\title{Wrapper for the deparse-substitute combo}
\usage{
desub(x)
}
\arguments{
\item{x}{object name}
}
\value{
character string
}
\description{
Transforms object name to character string.
}
\examples{
desub(trust)

}
