% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_se}
\alias{prop_se}
\title{Computes parametric standard error for proportions}
\usage{
prop_se(p, n)
}
\arguments{
\item{p}{proportion}

\item{n}{sample size}
}
\value{
numeric value representing SE for proportions
}
\description{
Computes SE for proportions using the old-school \eqn{\sqrt{\frac{p - (1 - p)}{n}}} formula.
}
\examples{
p = 0.3
n = 200
prop_se(p = p, n = n)
}
