% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_categorical_vars}
\alias{compose_spreadsheet}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{spreadview}{\code{\link[spreadview:compose_spreadsheet]{compose_spreadsheet()}}, \code{\link[spreadview:get_categorical_vars]{get_categorical_vars()}}}
}}

