% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se_mean}
\alias{se_mean}
\title{Computes parametric standard error for mean}
\usage{
se_mean(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric value representing SE for means
}
\description{
Computes SE for the mean using the old-school \eqn{\sqrt{\frac{var_x}{n}}} formula.
}
\examples{
se_mean(trust$age)
}
