% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se_median}
\alias{se_median}
\title{Computes parametric standard error for median}
\usage{
se_median(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric value representing SE for medians
}
\description{
Computes SE for the median using the old-school \eqn{\sqrt{1.2533\cdot\frac{var_x}{n}}} formula.
This assumes normal distribution.
}
\examples{
se_median(trust$age)
}
