% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{stem_chart_from}
\alias{stem_chart_from}
\title{Convert a plot unless it already is a chart}
\usage{
stem_chart_from(plot, chart_args = list())
}
\arguments{
\item{plot}{A ggplot or an \code{ms_chart}.}

\item{chart_args}{Arguments passed to \code{\link[=stem_as_mschart]{stem_as_mschart()}}.}
}
\value{
An \code{ms_chart}.
}
\description{
Convert a plot unless it already is a chart
}
\keyword{internal}
