% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_export.R
\name{stem_ggsave}
\alias{stem_ggsave}
\title{Export Stem plots}
\usage{
stem_ggsave(plot, name = NA, path = NULL, device = "png", dpi = 300, ...)
}
\arguments{
\item{plot}{Plot created by \verb{stem_plot_*}, e.g. \code{\link[=stem_plot_single]{stem_plot_single()}}.}

\item{name}{Name of exported file, which has to include extension (e.g. ".png"). If \code{NA}, will be the same as the plot object (extension not needed then).}

\item{path}{Optional path to export the plot to.}

\item{device}{Device to be used for exporting, e.g. "png" or "pdf".}

\item{dpi}{Resultion of the exported plot. Usually 300 for "medium" resolution, 600 for "high".}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
File with exported plot.
}
\description{
Wrapper around \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} which tries to export plots in a format appropriate for typical report.
Only works for plots created by \verb{stem_plot_*} family of functions.
}
\details{
This function aims at exporting common types of plots in correct size and resolution with minimal user input.
Each plot created by \verb{stem_plot_*} function has \code{plot_fn} attribute used to set default width and heigh of the output.
For example, plots created by \code{\link[=stem_plot_single]{stem_plot_single()}} have attribute value of \code{"plot_single"}, which leads to
them being exported in 16x4 cm dimensions.
}
\examples{
\dontrun{
species_plot <- stem_plot_single(iris, Species)

stem_ggsave(species_plot)
}

}
