% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{stem_palette_generator}
\alias{stem_palette_generator}
\title{Prepares Color Palette for ggplot2}
\usage{
stem_palette_generator(palette = "modern", direction = 1)
}
\arguments{
\item{palette}{Name of a palette.}

\item{direction}{If \code{-1}, reverses the order of colors.}
}
\value{
A vector of color hex codes to use with ggplot2
}
\description{
Prepares Color Palette for ggplot2
}
\examples{
\dontrun{
stem_palette_generator()
}
}
\keyword{internal}
