% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot}
\alias{stem_plot}
\title{General plotting function}
\usage{
stem_plot(
  data,
  item,
  group = NULL,
  weight = NULL,
  collapse_item = NULL,
  collapse_group = NULL,
  geom = ggplot2::geom_col,
  geom_args = list(position = ggplot2::position_dodge(width = 0.95)),
  title = TRUE,
  title_label = TRUE,
  title_wrap = 50,
  format_axis = TRUE,
  caption = FALSE,
  label = TRUE,
  label_args = list(position = ggplot2::position_dodge(width = 0.95), color = "black",
    vjust = -1),
  label_hide = 0.05,
  label_accuracy = 1,
  label_scale = 1,
  label_prefix = "",
  label_suffix = "",
  label_big = " ",
  label_decimal = ",",
  group_size = "none",
  group_wrap = NULL,
  item_size = "none",
  item_wrap = NULL,
  scale_x = ggplot2::scale_x_discrete(),
  scale_y = ggplot2::scale_y_continuous(limits = c(0, NA)),
  scale_fill = scale_fill_stem(palette = "modern"),
  scale_color = scale_color_stem(palette = "modern"),
  facet = FALSE,
  guides = ggplot2::guides(),
  coord_flip = FALSE,
  item_reverse = FALSE,
  group_reverse = FALSE
)
}
\arguments{
\item{data}{Dataframe including item (and group) variables}

\item{item}{Plotted item}

\item{group}{Optional; plotted grouping variable}

\item{weight}{Optional; survey weights}

\item{collapse_item}{Optional, a named list to collapse item categories}

\item{collapse_group}{Optional, a named list to collapse group categories}

\item{geom}{Geom that will represent data}

\item{geom_args}{Arguments for the geom function like size, color, etc.}

\item{title}{Should the name of item be used as plot title? Yes by default}

\item{title_label}{Should title use item label instead of name? Only if item has attribute "label".}

\item{title_wrap}{Length of title in characters before it gets wrapped. Defaults to 50.}

\item{format_axis}{If \code{TRUE}, adds title for x axis and removes titles for other scales.}

\item{caption}{If \code{TRUE}, adds a caption with sample size.}

\item{label}{Should geom labels be printed? Yes by default}

\item{label_args}{Arguments for the labels function like vjust, postion, etc.}

\item{label_hide}{Hides geom labels with values smaller than this threshold. Defaults to 0.05.}

\item{label_accuracy}{Accuracy geom labels. \code{1} means no decimal places, \code{0.1} one decimal place.}

\item{label_scale}{Multiplicative scale of geom labels. \code{100} multiple the values by 100}

\item{label_prefix}{Prefix for geom labels.}

\item{label_suffix}{Suffix for geom labels}

\item{label_big}{Character to split big numbers. Whitespace by default (e.g. \code{1000} will be printed as "1 000").}

\item{label_decimal}{Character to separate decimal digits. Comma by default (e.g. \code{1.2} will be printed as "1,2")}

\item{group_size}{Should the group size be included in the group legend? Either \code{n} (absolute frequency), \code{prop} (proportions) or \code{none} (nothing).}

\item{group_wrap}{How many characters in group categories before wrapping to the next line. Defaults to \code{NULL} - no wrapping.}

\item{item_size}{Should the group size be included in the item legend? Either \code{n} (absolute frequency), \code{prop} (proportions) or \code{none} (nothing).}

\item{item_wrap}{How many characters in item categories before wrapping to the next line. Defaults to \code{NULL} - no wrapping.}

\item{scale_x}{Pass \code{\link[ggplot2:scale_discrete]{ggplot2::scale_x_discrete()}} to the plot if needed.}

\item{scale_y}{Pass \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}} to the plot if needed.}

\item{scale_fill}{Pass \verb{scale_fill_*} to the plot if needed.}

\item{scale_color}{Pass \verb{scale_color_*} to the plot if needed.}

\item{facet}{Should groups be splitted into facets? No by default.}

\item{guides}{Pass \code{\link[ggplot2:guides]{ggplot2::guides()}} function to the plot if needed.}

\item{coord_flip}{If \code{TRUE}, switches x and y axes.}

\item{item_reverse}{Reverse order of item categories}

\item{group_reverse}{Reverse order of group categories}
}
\value{
a ggplot2 graph with custom attribute "stem_plot"
}
\description{
General plotting function
}
\examples{
stem_plot(data = trust,
item = government,
label = FALSE)

stem_plot(data = trust,
          item = police,
          group = eu_index,
          weight = W,
          geom_args = list(position = "stack"),
          label_args = list(position = ggplot2::position_stack(vjust = 0.5), color = "white"),
          label_scale = 100,
          label_suffix = "",
          item_size = "prop",
          group_size = "n",
          scale_y = ggplot2::scale_y_continuous(labels = scales::percent_format(accuracy = 1,
                                                                                suffix = " \%"))) +
theme_stem(legend.position = "bottom")
}
