% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_bar}
\alias{stem_plot_bar}
\title{Plot Item As Simple Barplot}
\usage{
stem_plot_bar(
  data,
  item,
  weight = FALSE,
  label = TRUE,
  nudge_label = 0.02,
  label_suffix = "\%",
  axis_suffix = "\%",
  axis_wrap = 40,
  reverse = FALSE,
  coord_flip = FALSE
)
}
\arguments{
\item{data}{Dataframe with the item to be plotted.}

\item{item}{Items to plot.}

\item{weight}{If \code{NA}, returns plot with unweighted frequencies. If (quoted) name of variable holding survey weights, returns weighted frequencies.}

\item{label}{Logical value. Should plot include numerical labels for each category?}

\item{nudge_label}{Distance between bar and label. Can be negative to put label inside bar.}

\item{label_suffix}{Suffix for labels inside the plot (e.g. "\%" or " \%").}

\item{axis_suffix}{Suffix for the x axis labels (e.g. "\%" or " \%").}

\item{axis_wrap}{Width of y axis label lines before the text gets wrapped.}

\item{reverse}{Reverses order of response categories.}

\item{coord_flip}{Reverses plot axes.}
}
\value{
A ggplot2 plot
}
\description{
Plots a single categorical variable as a simple barplot.
}
\examples{
stem_plot_bar(iris, Species)
}
