% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_barstack}
\alias{stem_plot_barstack}
\title{Plot a stacked barchart}
\usage{
stem_plot_barstack(
  data,
  item,
  group = NULL,
  weight = NULL,
  geom = ggplot2::geom_col,
  geom_args = list(position = "stack"),
  label_text_color = "black",
  label_args = list(position = ggplot2::position_stack(vjust = 0.5), color =
    label_text_color),
  scale_y = ggplot2::scale_y_continuous(labels = scales::percent_format(accuracy = 1,
    suffix = " \%"), limits = c(0, NA)),
  label = TRUE,
  label_scale = 100,
  label_hide = 0.05,
  coorf_flip = TRUE,
  item_reverse = TRUE,
  legend_rows = 1,
  legend_byrow = TRUE,
  palette = "modern",
  ...
)
}
\arguments{
\item{data}{Dataframe including item (and group) variables}

\item{item}{Plotted item}

\item{group}{Optional; plotted grouping variable}

\item{weight}{Optional; survey weights}

\item{geom}{geom that will represent data}

\item{geom_args}{Optional, a named list to collapse group categories}

\item{label_text_color}{Color of text labels. Default is \code{black}.}

\item{label_args}{Arguments for the labels function like vjust, postion, etc.}

\item{scale_y}{Format y axis using \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}}

\item{label}{Should geom labels be printed? Yes by default}

\item{label_scale}{Multiplicative scale of geom labels. \code{100} multiple the values by 100.}

\item{label_hide}{Hides geom labels with values smaller than this threshold. Defaults to 0.05.}

\item{coorf_flip}{Should the x and y axis be flipped? Defaults to \code{TRUE}}

\item{item_reverse}{Should the order of item categories be reversed? Defaults to \code{TRUE}}

\item{legend_rows}{Number of rows in legend}

\item{legend_byrow}{Shoul legend be filled by rows? (FALSE to fill by columns)}

\item{palette}{Stem palette to be used in the plot. See \code{\link[=scale_color_stem]{scale_color_stem()}} and \code{\link[=scale_fill_stem]{scale_fill_stem()}} for details.}

\item{...}{Other arguments passed to \code{\link[=stem_plot]{stem_plot()}}}
}
\value{
A ggplot2 object with custom attribute "stem_plot"
}
\description{
Wrapper around \code{\link[=stem_plot]{stem_plot()}} to make creating stacked barplots easier.
}
\examples{
stem_plot_barstack(trust, police, eu_index)
stem_plot_barstack(trust, police)

}
