% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_battery}
\alias{stem_plot_battery}
\title{Plot battery of like items}
\usage{
stem_plot_battery(
  data,
  items,
  weight = NULL,
  order_by = NULL,
  geom_args = list(position = "stack"),
  label_args = list(position = ggplot2::position_stack(vjust = 0.5), color =
    label_text_color),
  label_scale = 100,
  label_text_color = "black",
  label_hide = 0.05,
  item_label = TRUE,
  group_size = "none",
  scale_y = ggplot2::scale_y_continuous(labels = scales::percent_format(accuracy = 1,
    suffix = " \%")),
  coord_flip = TRUE,
  palette = "div1",
  legend_rows = 1,
  legend_byrow = TRUE,
  item_reverse = TRUE,
  group_reverse = TRUE,
  ...
)
}
\arguments{
\item{data}{Dataframe including item (and group) variables}

\item{items}{Plotted items. Can be selected using Tidyselect's selection helpers.}

\item{weight}{Optional; survey weights}

\item{order_by}{Vector of response categories to order items by, e.g. \code{c("Definitely Agree", "Rather Agree")}.}

\item{geom_args}{Arguments for the geom function like size, color, etc.}

\item{label_args}{Arguments for the labels function like vjust, postion, etc.}

\item{label_scale}{Multiplicative scale of geom labels. \code{100} multiple the values by 100.}

\item{label_text_color}{Color of text labels. Default is \code{black}.}

\item{label_hide}{Hides geom labels with values smaller than this threshold. Defaults to 0.05.}

\item{item_label}{If \code{TRUE}, item labels are used instead of names (all items must have a label).}

\item{group_size}{Should the group size be included in the group legend? Either \code{n} (absolute frequency), \code{prop} (proportions) or \code{none} (nothing).}

\item{scale_y}{Format y axis using \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}}

\item{coord_flip}{If \code{TRUE}, switches x and y axes.}

\item{palette}{Stem palette to be used in the plot. See \code{\link[=scale_color_stem]{scale_color_stem()}} and \code{\link[=scale_fill_stem]{scale_fill_stem()}} for details.}

\item{legend_rows}{Number of rows in legend}

\item{legend_byrow}{Shoul legend be filled by rows? (FALSE to fill by columns)}

\item{item_reverse}{Should the order of item categories be reversed? Defaults to \code{TRUE}}

\item{group_reverse}{Should the order of group categories be reversed? Defaults to \code{TRUE}}

\item{...}{Other arguments passed to \code{\link[=stem_plot]{stem_plot()}}}
}
\value{
A ggplot2 object with custom attribute "stem_plot"
}
\description{
Plots a set of categorical variables with same response categories as a single graph.
}
\examples{
\dontrun{
stem_plot_battery(trust,
                  items = c(police, eu, government, army),
                  weight = W,
                  order_by = c("Definitely Agree", "Rather Agree"))
}
}
