% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_freq_bar}
\alias{stem_plot_freq_bar}
\title{Plot Frequency of category by another variable}
\usage{
stem_plot_freq_bar(
  data,
  item,
  group = NA,
  weight = FALSE,
  freq_by,
  axis_suffix = "\%",
  axis_wrap = 20,
  nudge_label = 0.02,
  segment_size = 4,
  segment_suffix = "",
  order_group = TRUE,
  reverse = FALSE,
  coord_flip = FALSE
)
}
\arguments{
\item{data}{Dataframe with the items to be plotted.}

\item{item}{Item to be aggregated and ploted.}

\item{group}{Grouping variable.}

\item{weight}{Optional suvey weights.}

\item{freq_by}{Character vector of responses whose frequencies are ploted.}

\item{axis_suffix}{Axis label suffix (e.g. "\%").}

\item{axis_wrap}{Number of Axis labels characters before the text is wrapped.}

\item{nudge_label}{Numeric, distance between bar and label. Can be negative to put label inside bar.}

\item{segment_size}{Font size for segments labels.}

\item{segment_suffix}{Suffix for labels inside the plot (e.g. "\%" or " \%").}

\item{order_group}{Logical, should the grouping variable be ordered by frequencies?}

\item{reverse}{Logical, should the order of grouping categories be reversed?}

\item{coord_flip}{Logical, should the axis be reversed?}
}
\value{
A ggplot object
}
\description{
Aggregates category frequencies and plots them based on grouping variable.
}
