% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{stem_plot_list}
\alias{stem_plot_list}
\title{Collect plots passed through \code{...}}
\usage{
stem_plot_list(...)
}
\arguments{
\item{...}{Plots, or one list of plots.}
}
\value{
A list of plots.
}
\description{
Internal helper accepting either several plots in \code{...} or a single list of
plots.
}
\keyword{internal}
