% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_multiple}
\alias{stem_plot_multiple}
\title{Plot Multiple Items in Line}
\usage{
stem_plot_multiple(
  data,
  items,
  weight = NA,
  var_labels = NA,
  order_by = NA,
  reverse_responses = TRUE,
  labels = TRUE,
  hide_labels = 0.05,
  labels_size = 5,
  labels_color = "black",
  labels_suffix = "\%",
  axis_suffix = "\%",
  axis_wrap = 80
)
}
\arguments{
\item{data}{Dataframe with the items to be plotted.}

\item{items}{Items to plot, can be selected using dplyr's selection helpers.}

\item{weight}{If \code{NA}, returns plot with unweighted frequencies. If (quoted) name of variable holding survey weights, returns weighted frequencies.}

\item{var_labels}{Optional dataframe with item labels used in the plot. Has to include columns \code{item} and \code{label}.}

\item{order_by}{Optional vector of response category/categories. Items in the plot will be sorted based on their frequencies.}

\item{reverse_responses}{Reverses order of response categories. Defaults to \code{TRUE}, since most often we want positive responses on the left.}

\item{labels}{Logical value. Should plot include numerical labels for each category?}

\item{hide_labels}{Value between 0 and 1. Hides labels with with frequencies lower than this value.}

\item{labels_size}{Font size for labels inside the plot.}

\item{labels_color}{Color of labels inside the plot. Can be either label ("black", "white") or hexcode.}

\item{labels_suffix}{Suffix for labels inside the plot (e.g. "\%" or " \%").}

\item{axis_suffix}{Suffix for the x axis labels (e.g. "\%" or " \%").}

\item{axis_wrap}{Width of y axis label line before the text gets wrapped.}
}
\value{
A ggplot2 graph.
}
\description{
Plot multiple categorical items (e.g. battery of likert items) in a stacked barplot.
}
