% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_scatter}
\alias{stem_plot_scatter}
\title{Plot a simple scatter chart}
\usage{
stem_plot_scatter(
  data,
  x,
  y,
  group = NULL,
  weight = NULL,
  scale_x = ggplot2::scale_x_continuous(),
  scale_y = ggplot2::scale_y_continuous(),
  label = TRUE,
  label_scale = 1,
  label_hide = 0.05
)
}
\arguments{
\item{data}{Dataframe including items for x and y axes}

\item{x}{Plotted item on the x-axis}

\item{y}{Plotted item on the y-axis}

\item{group}{Optional; plotted grouping variable}

\item{weight}{Optional; survey weights}

\item{scale_x}{Optional; scaling for the x-axis}

\item{scale_y}{Optional; scaling for the y-axis}

\item{label}{Should geom labels be printed? Yes by default}

\item{label_scale}{Multiplicative scale of geom labels}

\item{label_hide}{Threshold below which labels are hidden}
}
\value{
A ggplot2 object
}
\description{
This function creates a scatter plot directly using ggplot2. It allows for customization of plot aesthetics and includes options for weighting and grouping data.
}
\examples{
stem_plot_scatter(data = trust,x = age,y = W)
}
