% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_single}
\alias{stem_plot_single}
\title{Plot a Single Item in Line}
\usage{
stem_plot_single(
  data,
  item,
  weight = FALSE,
  wrap_width = 17,
  stem_palette = "modern",
  text_color = "black",
  axis_suffix = "\%",
  segment_suffix = "\%"
)
}
\arguments{
\item{data}{Dataframe with the item to be plotted.}

\item{item}{Item to plot}

\item{weight}{If \code{FALSE} raw frequencies are ploted. Can be a variable holding survey weights. For now, the name of the variable has to be quoted.}

\item{wrap_width}{Length of the legend categories names before they are wrapped to the next line.}

\item{stem_palette}{Name of a palette from \code{\link[=stem_palettes]{stem_palettes()}}.}

\item{text_color}{Color of the labels inside the plot.}

\item{axis_suffix}{Suffixes of labels on the X axis.}

\item{segment_suffix}{Suffixes of labels inside the plot.}
}
\value{
A ggplot2 graph.
}
\description{
Plots a single categorical item (usually a likert item) as a horizontal barplot.
}
\examples{
stem_plot_single(iris, Species)
}
