% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{stem_summarise}
\alias{stem_summarise}
\title{Summarise numeric or categorical variable (possibly segmented)}
\usage{
stem_summarise(
  data,
  item,
  group = NULL,
  weight = NULL,
  long = FALSE,
  collapse_item = NULL,
  collapse_group = NULL,
  return_n = FALSE
)
}
\arguments{
\item{data}{Dataframe including variables to be analyzed}

\item{item}{Variable to be summarised}

\item{group}{Optional segmenting variable}

\item{weight}{Optional survey weights}

\item{long}{Returns data in long format. Useful if multiple dataframes are to be merged}

\item{collapse_item}{Named list. Optionally collapes (or renames) categories of the item variable (Only if item is categorical).}

\item{collapse_group}{Named list. Optionally collapes (or renames) categories of the group variable}

\item{return_n}{If \code{TRUE}, returns absolute group sizes.}
}
\value{
An aggregated data frame with point estimates (either proportions or means) and 95\% confidence intervals.
}
\description{
Summarise either numeric or categerical variable, possible segment by another categorical variable.
Actually a wrapper for \code{\link[=stem_summarise_cat]{stem_summarise_cat()}} and \code{\link[=stem_summarise_num]{stem_summarise_num()}}.
}
\details{
Apart from either point estimates (proportions or means), the function also returns 95\% confidence interval bounds.
If unweighted, the intervals are computed using the basic \code{sqrt((p * (1-p)) / n)} formula. If the estimated proportions
are very high/low, this may lead to interval estimates outsides of the (0;1) bounds. If weights are used, the confidence
intervals are based on weighted logistic regression. Neither of the approaches will work if the proportions are exactly zero or one.

If \code{long = TRUE}, new column is added holding name of the item (and group) variable. This is useful if you need to loop through
multiple variables and bind the results into a single data frame. See online vignettes for details.

\code{collapse_item} and \code{collapse_group} can be used to collapse categories of item and group variable. The named list should include
vector of old category names with corresponding to the new category name, e.g. \verb{list(Agree = c("Definitely Agree", "Rather Agree")}.
You can also use the arguments to rename categories (\code{list(Yes = "Agree")}) or pass other arguments from the \code{\link[forcats:fct_collapse]{forcats::fct_collapse()}} function.

If \code{return_n = TRUE}, columns holding the absolute frequencies will be added. Column \code{n} is the number of observations for the specific combination of
item and grouping variable, \code{n_item} is the frequency of the item categories and \code{n_group} is the frequency of the group categories.
}
\examples{
stem_summarise(data = trust, item = government)

stem_summarise(data = trust, item = age, group = eu_index)

stem_summarise(data = trust, item = government,
               collapse_item = list(Agree = c("Definitely Agree", "Rather Agree")))

}
