% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{stem_summarise_num}
\alias{stem_summarise_num}
\title{Summarise numeric variable using arithmetic mean (possibly segmented)}
\usage{
stem_summarise_num(
  data,
  item,
  group = NULL,
  weight = NULL,
  long = FALSE,
  collapse_group = NULL,
  return_n = FALSE
)
}
\arguments{
\item{data}{Dataframe including data to be analyzed}

\item{item}{Variable to be summarised}

\item{group}{Optional segmenting variable}

\item{weight}{Optional survey weights}

\item{long}{Returns data in long format. Useful if multiple dataframes are to be merged}

\item{collapse_group}{Named list. Optionally collapses (or renames) categories of the group variable}

\item{return_n}{If \code{TRUE}, returns absolute group sizes.}
}
\value{
Tidy dataframe
}
\description{
Takes one numeric variable and returns a tidy dataframe, including relative mean and
absolute size of each group. The variable can be further segmented by another categorical variable.
Can also take survey weights. Returns 95\% confidence intervals.
}
\examples{
trust |> stem_summarise_num(age,
                            group = eu_index,
                            weight = W,
         collapse_group = list(`Neutral or no opinion` = c("Neutral", "Doesn't Know")))
}
