% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-project.R
\name{create_project}
\alias{create_project}
\title{This package will create a function called create_project()}
\usage{
create_project(path, ...)
}
\arguments{
\item{path}{Path to the project}

\item{...}{Additional arguments, used to collect user input}
}
\description{
It's callback is at: inst/rstudio/templates/project/create_project.dcf
}
