% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{scale_colour_stem}
\alias{scale_colour_stem}
\alias{scale_color_stem}
\alias{scale_fill_stem}
\title{Sequential, diverging and qualitative color scales from \code{stemtools}}
\usage{
scale_colour_stem(palette = "modern", direction = 1, ...)

scale_color_stem(palette = "modern", direction = 1, ...)

scale_fill_stem(palette = "modern", direction = 1, ...)
}
\arguments{
\item{palette}{Name of a palette to use.}

\item{direction}{Orientation of the palette. If \code{-1}, the order of colors is reversed.}

\item{...}{Additional arguments to be passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\description{
Returns a vector of colors forming a cohesive palette. Current palettes include:
\itemize{
\item Nominal: \code{gruvbox}
\item Diverging: \code{gruvbox_div}, \code{modern}
}
}
\examples{
ggplot2::ggplot(data = mtcars,
       mapping = ggplot2::aes(x = mpg, y = hp, color = as.factor(am))) +
       ggplot2::geom_point() +
       scale_color_stem()
}
