% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_bar}
\alias{stem_plot_bar}
\title{Plot a simple barchart}
\usage{
stem_plot_bar(
  data,
  item,
  group = NULL,
  weight = NULL,
  scale_y = ggplot2::scale_y_continuous(labels = scales::percent_format(accuracy = 1),
    limits = c(0, NA)),
  label = TRUE,
  label_scale = 100,
  label_hide = 0,
  ...
)
}
\arguments{
\item{data}{Dataframe including item (and group) variables}

\item{item}{Plotted item}

\item{group}{Optional; plotted grouping variable}

\item{weight}{Optional; survey weights}

\item{scale_y}{Format y axis using \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}}

\item{label}{Should geom labels be printed? Yes by default}

\item{label_scale}{Multiplicative scale of geom labels. \code{100} multiple the values by 100}

\item{label_hide}{Hides geom labels with values smaller than this threshold. Defaults to 0.}

\item{...}{Other arguments passed to \code{\link[=stem_plot]{stem_plot()}}}
}
\value{
A ggplot2 object with custom attribute "stem_plot"
}
\description{
Wrapper around \code{\link[=stem_plot]{stem_plot()}} to make creating barplots easier.
}
\examples{
stem_plot_bar(trust, government)
}
