% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_bar}
\alias{stem_plot_bar}
\title{Plot Item As Simple Barplot}
\usage{
stem_plot_bar(
  data,
  item,
  group = NA,
  weight = FALSE,
  label = TRUE,
  nudge_label = 0.02,
  segment_size = 4,
  segment_suffix = "",
  hide_labels = 0.05,
  axis_suffix = "\%",
  axis_wrap = 40,
  axis_expand = ggplot2::waiver(),
  reverse_responses = TRUE,
  reverse_groups = FALSE,
  reverse_legend = TRUE,
  order_by = NA,
  coord_flip = FALSE
)
}
\arguments{
\item{data}{Dataframe with the item to be plotted.}

\item{item}{Items to plot.}

\item{group}{Optional grouping variable}

\item{weight}{If \code{NA}, returns plot with unweighted frequencies. If name of variable holding survey weights, returns weighted frequencies.}

\item{label}{Logical value. Should plot include numerical labels for each category?}

\item{nudge_label}{Distance between bar and label. Can be negative to put label inside bar.}

\item{segment_size}{Font size for segments labels.}

\item{segment_suffix}{Suffix for labels inside the plot (e.g. "\%" or " \%").}

\item{hide_labels}{Number between 0 and 1. Labels with smaller frequencies are hidden.}

\item{axis_suffix}{Suffix for the x axis labels (e.g. "\%" or " \%").}

\item{axis_wrap}{Width of y axis label lines before the text gets wrapped.}

\item{axis_expand}{Vector of coordinates to expand axis with frequencies.}

\item{reverse_responses}{Reverses order of response categories.}

\item{reverse_groups}{Reverse order of grouping variable categories.}

\item{reverse_legend}{If TRUE, reverses order of response categories in legend.}

\item{order_by}{In grouped plot, categories to order item by.}

\item{coord_flip}{Reverses plot axes.}
}
\value{
A ggplot2 plot
}
\description{
Plots a single categorical variable as a simple barplot.
}
\examples{
stem_plot_bar(iris, Species)
}
