% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stem_plot_bartable}
\alias{stem_plot_bartable}
\title{Plot Table of Barplots}
\usage{
stem_plot_bartable(
  data,
  row,
  column,
  item,
  weight = FALSE,
  stem_palette = "modern",
  reverse_responses = TRUE,
  reverse_legend = TRUE,
  reverse_row = FALSE,
  reverse_column = FALSE,
  label = TRUE,
  hide_labels = 0.05,
  segment_size = 4,
  segment_suffix = "",
  wrap_row = 20,
  wrap_col = 20
)
}
\arguments{
\item{data}{Dataframe with the items to be plotted.}

\item{row}{Item in table row.}

\item{column}{Item in table column.}

\item{item}{Item inside the table.}

\item{weight}{If \code{NA}, returns plot with unweighted frequencies. If (quoted) name of variable holding survey weights, returns weighted frequencies.}

\item{stem_palette}{Name of a palette from \code{\link[=stem_palettes]{stem_palettes()}}.}

\item{reverse_responses}{Reverses order of response categories.}

\item{reverse_legend}{If TRUE, reverses order of response categories in legend.}

\item{reverse_row}{Reverses response categories of the row variable.}

\item{reverse_column}{Reverses response categories of the column variable.}

\item{label}{Logical value. Should plot include numerical labels for each category?}

\item{hide_labels}{Value between 0 and 1. Hides labels with with frequencies lower than this value.}

\item{segment_size}{Font size for segments labels.}

\item{segment_suffix}{Suffix for labels inside the plot (e.g. "\%" or " \%").}

\item{wrap_row}{Width of row label lines before the text gets wrapped.}

\item{wrap_col}{Width of column label lines before the text gets wrapped.}
}
\value{
A ggplot2 graph
}
\description{
Plots three categorical variables, one represented by table rows, second by table columns and the last one by a horizontal barplot.
}
