% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_export.R
\name{theme_stem}
\alias{theme_stem}
\title{Basic ggplot2 theme for Stem.}
\usage{
theme_stem(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
ggplot2 \code{\link[ggplot2:theme]{ggplot2::theme()}} object.
}
\description{
Basic ggplot2 theme for Stem.
}
\examples{
stem_plot_single(iris, Species) + theme_stem()
}
