% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trust}
\alias{trust}
\title{Simulated data on attitudes towards institutions}
\format{
\subsection{\code{trust}}{

A data frame with 1 000 rows and 9 columns:
\describe{
\item{police}{Generally speaking, can the following be trusted: Police}
\item{government}{Generally speaking, can the following be trusted: Czech government}
\item{eu}{Generally speaking, can the following be trusted: European Union}
\item{army}{Generally speaking, can the following be trusted: Army}
\item{scientists}{Generally speaking, can the following be trusted: Scientists}
\item{eu_index}{Attitudes towards EU index}
\item{nat_index}{Openness towards society index}
\item{age}{Age}
\item{W}{Sampling weight}
...
}
}
}
\usage{
trust
}
\description{
Simulated data to use for examples and function testing.
}
\keyword{datasets}
